requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.4
django-environ==0.11.2
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
pypng==0.20220715.0
python-barcode==0.15.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
setuptools==69.5.1
simplejson==3.19.2
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
